fastapi/search/requirements-search.txt
annotated-types
anyio
uvicorn
certifi
cffi
charset-normalizer
colorama
cryptography
Deprecated
distro
elastic-transport
elasticsearch
fastapi
filelock
regex
fsspec
h11
httpcore
httpx
huggingface-hub
idna
Jinja2
jiter
MarkupSafe
mpmath
konlpy
networkx
numpy
openai
packaging
pycparser
pydantic
pydantic_core
PyGithub
PyJWT
PyNaCl
python-dotenv
PyYAML
regex
requests
safetensors
setuptools
sniffio
starlette
sympy
tokenizers
torch
tqdm
transformers
typing_extensions
urllib3
wrapt
